+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 ## [Unreleased]
 ### Added
 - [PHP, Java, Ruby, JavaScript] update dependency messages up to v26
+- [Python] Added type annotations ([#283](https://github.com/cucumber/gherkin/pull/283))
 
 ### Changed
 - [.NET] Drop unsupported frameworks. Now supported target frameworks are .NET 8, .NET Standard 2.0 ([#265](https://github.com/cucumber/gherkin/pull/265))